Information Window
Created a unique window for each scenario in which a user needs more
information in order to do a calculation for a particular chemical. The
PvtWindow (which should be renamed) is practically finished. All that
remains in the stoic portion of the program is to link the different
sections of calculations together. Also added an exit button and a back
to the main screen button. (also we've reached our 1000th line of code
in the GUI file, woo!)